+++ packages/backend/src/controllers/demandeSejour/getExtractFO.js
@@ -5,6 +5,7 @@ const { escapeCsvField } = require("../../utils/csv");
 const dayjs = require("dayjs");
 const logger = require("../../utils/logger");
 const AppError = require("../../utils/error");
+const { formatSiret } = require("../../utils/siret");
 
 const log = logger(module.filename);
 
@@ -46,6 +47,7 @@ module.exports = async function get(req, res, next) {
       "libelle",
       "departement",
       "etablissement",
+      "siret",
       "date_debut",
       "date_fin",
       "statut",
@@ -57,6 +59,7 @@ module.exports = async function get(req, res, next) {
         const newItem = { ...item };
         newItem.date_debut = dayjs(item.date_debut).format("DD/MM/YYYY");
         newItem.date_fin = dayjs(item.date_fin).format("DD/MM/YYYY");
+        newItem.siret = formatSiret(newItem.siret);
         newItem.etablissement =
           item.type_organisme === "personne_morale" ? item.raison_sociale : "";
         return [
