+++ packages/backend/src/controllers/organisme/getListeExtract.js
@@ -3,6 +3,7 @@ const Organisme = require("../../services/Organisme");
 
 const logger = require("../../utils/logger");
 const dayjs = require("dayjs");
+const { formatSiren, formatSiret } = require("../../utils/siret");
 
 const log = logger(module.filename);
 
@@ -110,25 +111,3 @@ module.exports = async function get(_req, res, next) {
     return next(error);
   }
 };
-function formatByPattern(pattern, input = "", separator = " ") {
-  if (typeof input !== "string") input = String(input ?? "");
-  const expectedLength = pattern.reduce((a, b) => a + b, 0);
-  if (input.length !== expectedLength || !/^\d+$/.test(input)) return input;
-
-  let i = 0;
-  return pattern
-    .map((len) => {
-      const part = input.slice(i, i + len);
-      i += len;
-      return part;
-    })
-    .join(separator);
-}
-
-function formatSiret(s) {
-  return formatByPattern([3, 3, 3, 5], s);
-}
-
-function formatSiren(s) {
-  return formatByPattern([3, 3, 3], s);
-}
