+++ packages/backend/src/services/DemandeSejour.js
@@ -738,6 +738,7 @@ WHERE uo.org_id = $1 AND u.status_code = 'VALIDATED'
       ds.date_debut::text as "date_debut",
       ds.date_fin::text as "date_fin",
       pm.raison_sociale as "raison_sociale",
+      STRING_AGG(DISTINCT COALESCE(pm.siret, pp.siret), ', ') AS siret,
       (
           SELECT adr.departement
           FROM front.demande_sejour_to_hebergement dsth
@@ -762,6 +763,14 @@ WHERE uo.org_id = $1 AND u.status_code = 'VALIDATED'
     LEFT JOIN front.personne_physique pp ON pp.organisme_id = o.id
     WHERE
       o.id = ANY ($1)
+    GROUP BY ds.id,
+      ds.statut,
+      ds.id_fonctionnelle,
+      o.type_organisme ,
+      ds.libelle ,
+      ds.date_debut,
+      ds.date_fin,
+      pm.raison_sociale
     ORDER BY date_debut
     `,
   getHebergementsByDepartementCodes: () => `
